feat: Created new analyzer.
The new analyser works on `manifest.trivy` files (name up for debate!)
and adds packages from it.
The manifest file is a json file with a single array, each object in the
array can contain any data which the types.Package object have, only
name and version is _required_ though.

Signed-off-by: Johannes Tegnér <johannes@jitesoft.com>